requirements.txt
# Python dependencies for plate violation processor
# متطلبات Python لمعالج المخالفات

# Plate Recognition API
requests>=2.31.0

# PostgreSQL Database Driver
psycopg2-binary>=2.9.9

# PDF Generation
fpdf>=1.7.2

# Excel Generation
pandas>=2.0.0
openpyxl>=3.1.0

# Image Processing
Pillow>=10.0.0
